Fix attempt for non-ascii escape characters.
The fundamental problem is that the bytes::escaped_transform still a
control_char instead of a byte or byte sequence. What is important here
is that it breaks for control characters which have a multi-byte
encoding (i.e. non-ascii byte values) but still represent single bytes.
For protocols using those, bytes::escaped_transform can not be used.

Keep in mind that this change may affect non-ascii unicode separators
negatively.